ch(endpoints):Following conventions -removed all the user validating codes from controllers -added some validating codes in the auth.middleware -deleted to some folders which should be include in .gitignore [finishes #169950991]
