+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"22eeb853-c4a6-4d2f-a200-3421c9037df3":{"parent":null,"children":[]}},"files":{"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\server.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\auth.route.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\users.controller.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\user.model.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\data.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\redflag.model.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\auth.middleware.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\redflags.route.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\redflags.controller.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\token.middleware.js":["22eeb853-c4a6-4d2f-a200-3421c9037df3"]},"externalIds":{}}
+{"processes":{"d07ff308-19fb-43cc-b739-412c9f90ec73":{"parent":null,"children":[]}},"files":{"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\server.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\auth.route.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\users.controller.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\user.model.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\data.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\redflag.model.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\auth.middleware.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\redflags.route.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\redflags.controller.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\token.middleware.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"]},"externalIds":{}}
